Refactor git namer into separate files with additional tests
